+++ app.py
@@ -37,6 +37,7 @@ from lastfm.routes import lastfm_blueprint
 from discogs.routes import discogs_blueprint
 from cards.routes import cards_blueprint
 from llm.routes import llm_blueprint
+from admin.routes import admin_blueprint
 
 app.register_blueprint(albums_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/a')
 app.register_blueprint(racks_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/r')
@@ -45,3 +46,4 @@ app.register_blueprint(lastfm_blueprint, url_prefix=f'{app.config["API_PREFIX"]}
 app.register_blueprint(discogs_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/discogs')
 app.register_blueprint(cards_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/card')
 app.register_blueprint(llm_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/llm')
+app.register_blueprint(admin_blueprint, url_prefix=f'{app.config["API_PREFIX"]}/admin')